# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JCEmpleos API",
        "description": "Documentación oficial del API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/perfil/educacion": {
            "get": {
                "tags": [
                    "Perfil - Educación"
                ],
                "summary": "Listar formación académica del usuario",
                "operationId": "b2f82c9dc6b56e461191f582078555e6",
                "responses": {
                    "200": {
                        "description": "Listado obtenido correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Perfil - Educación"
                ],
                "summary": "Agregar formación académica",
                "operationId": "085a55a46201c2778eb2db529231ce04",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "institucion"
                                ],
                                "properties": {
                                    "institucion": {
                                        "type": "string",
                                        "example": "Universidad Nacional"
                                    },
                                    "nivel_educativo_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "area_estudio_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "fecha_desde": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2015-01-01"
                                    },
                                    "fecha_hasta": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-12-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Educación agregada correctamente"
                    },
                    "422": {
                        "description": "Error al agregar educación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/perfil/educacion/{id}": {
            "put": {
                "tags": [
                    "Perfil - Educación"
                ],
                "summary": "Actualizar formación académica",
                "operationId": "056a4dc9a786836413f484abdf025b2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "institucion": {
                                        "type": "string"
                                    },
                                    "nivel_educativo_id": {
                                        "type": "integer"
                                    },
                                    "area_estudio_id": {
                                        "type": "integer"
                                    },
                                    "fecha_desde": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "fecha_hasta": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Educación actualizada correctamente"
                    },
                    "422": {
                        "description": "Error al actualizar educación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Perfil - Educación"
                ],
                "summary": "Eliminar formación académica",
                "operationId": "318f0b10a4b96059420eeba2c1c5cea8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Educación eliminada correctamente"
                    },
                    "422": {
                        "description": "Error al eliminar educación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}